Fixed issue caused by port from 1.20.1

+++ src/main/java/team/creative/creativecore/common/gui/controls/simple/GuiIcon.java
@@ -5,8 +5,8 @@ import com.mojang.blaze3d.vertex.PoseStack;
 
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.GameRenderer;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
 import team.creative.creativecore.client.render.GuiRenderHelper;
 import team.creative.creativecore.common.gui.GuiChildControl;
 import team.creative.creativecore.common.gui.GuiControl;
@@ -30,39 +30,39 @@ public class GuiIcon extends GuiControl {
         this.shadow = Color.NONE;
         this.color = Color.WHITE;
     }
-
+    
     public GuiIcon setIcon(Icon icon) {
         this.icon = icon;
         return this;
     }
-
+    
     public GuiIcon setColor(Color color) {
         this.color = color;
         return this;
     }
-
+    
     public GuiIcon setShadow(Color shadowColor) {
         this.shadow = shadowColor;
         return this;
     }
-
+    
     public GuiIcon setSquared(boolean squared) {
         this.squared = squared;
         return this;
     }
-
+    
     @Override
     public GuiIcon setDim(int width, int height) {
         super.setDim(width, height);
         return this;
     }
-
+    
     @Override
     public GuiIcon setDim(GuiSizeRule dim) {
         super.setDim(dim);
         return this;
     }
-
+    
     @Override
     public void init() {
         
@@ -113,25 +113,26 @@ public class GuiIcon extends GuiControl {
         RenderSystem.defaultBlendFunc();
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
         RenderSystem.setShaderTexture(0, this.icon.location());
-
+        
         int x = 0, y = 0, width = control.getContentWidth(), height = control.getContentHeight();
         if (squared) {
             int size = Math.min(width, height);
             int diff = Math.abs(width - height);
-            if (width == size) y += diff / 2;
-            else x += diff / 2;
+            if (width == size)
+                y += diff / 2;
+            else
+                x += diff / 2;
             width = height = size;
         }
-
+        
         if (this.shadow != Color.NONE) {
             this.shadow.glColor();
-            GuiRenderHelper.textureRect(pose, x + 1, y + 1, width, height, (float) this.icon.minX(), (float) this.icon.minY(), (float) (this.icon
-                    .minX() + this.icon.width()), (float) (this.icon.minY() + this.icon.height()));
+            GuiRenderHelper.textureRect(pose, x + 1, y + 1, width, height, this.icon.minX(), this.icon.minY(), this.icon.minX() + this.icon.width(), this.icon.minY() + this.icon
+                    .height());
         }
-
+        
         this.color.glColor();
-        GuiRenderHelper.textureRect(pose, x, y, width, height, (float) this.icon.minX(), (float) this.icon.minY(), (float) (this.icon
-                .minX() + this.icon.width()), (float) (this.icon.minY() + this.icon.height()));
+        GuiRenderHelper.textureRect(pose, x, y, width, height, this.icon.minX(), this.icon.minY(), this.icon.minX() + this.icon.width(), this.icon.minY() + this.icon.height());
         RenderSystem.disableBlend();
         pose.popPose();
     }
